pyproject.toml
[project]
name = "ai-chaos-theory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "h5py>=3.15.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "phiflow>=3.4.0",
    "scipy>=1.16.3",
    "torch>=2.9.1",
]
